add authenticationFacade for userbean and make update data check by role, update comment only can update or delete own self comment

+++ src/main/java/com/example/tickettrackingsystem/controller/BugController.java
@@ -1,20 +1,20 @@
 package com.example.tickettrackingsystem.controller;
 
 import com.example.tickettrackingsystem.dao.BugDao;
-import com.example.tickettrackingsystem.model.UserDto;
 import com.example.tickettrackingsystem.model.tracking.Bug;
 import com.example.tickettrackingsystem.service.BugService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Map;
+
 
 @CrossOrigin(origins = "*", maxAge = 3600)
 @RestController
-@RequestMapping("/bug")
+@RequestMapping("/api")
 public class BugController {
 
     @Autowired
@@ -24,48 +24,58 @@ public class BugController {
     private BugDao bugDao;
 
     @PreAuthorize("hasRole('QA') || hasRole('ADMIN')")
-    @RequestMapping(value="/", method = RequestMethod.POST)
+    @PostMapping(value="/bug")
     @ResponseStatus( HttpStatus.CREATED )
-    public Bug createBug(@RequestBody Bug bug){
-        bug.setResolved(false);
-        return bugDao.save(bug);
-    }
-
-    @PreAuthorize("hasRole('QA') || hasRole('ADMIN')")
-    @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-    public ResponseEntity<Object> deleteBugById(@PathVariable Long id) {
+    public ResponseEntity<Object> createBug(@RequestBody Bug bug){
         try {
-            bugDao.deleteById(id);
-        } catch (EmptyResultDataAccessException emptyResultDataAccessException) {
-            return new ResponseEntity<>("No Bug issue exist for id: " + id, HttpStatus.GONE);
+            bug = bugService.createBug(bug);
+        } catch (Exception e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
         }
-        return new ResponseEntity<>("Succeed", HttpStatus.OK);
+        return new ResponseEntity<>(bug, HttpStatus.CREATED);
     }
 
     @PreAuthorize("hasRole('QA') || hasRole('ADMIN')")
-    @RequestMapping(value="/{id}", method = RequestMethod.PUT)
-    public Bug updateBugById(@RequestBody Bug bug, @PathVariable Long id){
-        return bugService.updateBugById(id, bug);
+    @DeleteMapping(value = "/bug/{id}")
+    public ResponseEntity<Object> deleteBugById(@PathVariable Long id) {
+        if (!bugDao.existsById(id)) {
+            return new ResponseEntity<>(HttpStatus.GONE);
+        }
+        bugDao.deleteById(id);
+
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    @PreAuthorize("hasRole('RD') || hasRole('ADMIN')")
-    @RequestMapping(value="/{id}", method = RequestMethod.PATCH)
-    public Bug patchBugById(@RequestBody Bug bug, @PathVariable Long id){
-        return bugService.patchBugById(id, bug);
+    @PreAuthorize("hasRole('QA') || hasRole('ADMIN') || hasRole('RD')")
+    @PutMapping(value="/bug/{id}")
+    public ResponseEntity<Object> updateBugById(@RequestBody Bug bug, @PathVariable Long id){
+        if (!bugDao.existsById(id)) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(bugService.updateBugByRole(id, bug), HttpStatus.OK);
     }
 
     @PreAuthorize("hasRole('USER')")
-    @RequestMapping(value="/", method = RequestMethod.GET)
-    public Iterable<Bug> listBug(){
-        return bugDao.findAll();
+    @GetMapping("/bug")
+    public ResponseEntity<Map<String, Object>> getAllBugsPageable(@RequestParam(defaultValue = "0") int page,
+                                                                  @RequestParam(defaultValue = "3") int size) {
+        try {
+            Map<String, Object> response = bugService.getPageableResultByPageAndSize(page, size);
+            return new ResponseEntity<>(response, HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @PreAuthorize("hasRole('USER')")
-    @RequestMapping(value="/{id}", method = RequestMethod.GET)
-    public Bug getBugById(@PathVariable Long id){
-        return bugDao.findBugById(id);
+    @GetMapping(value="/{id}")
+    public ResponseEntity<Object> getBugById(@PathVariable Long id){
+        if (!bugDao.existsById(id)) {
+            return new ResponseEntity<>("Not found Bug with id: " + id, HttpStatus.NOT_FOUND);
+        }
+        Bug bug = bugDao.findBugById(id);
+        return new ResponseEntity<>(bug, HttpStatus.OK);
     }
-
 
 
 }
